Fix obsolete/broken git grep code
`values` causes a null symbol function error. I vaguely remembered
that it just returned the values as a list anyway, and they were
treated as a list at the call site, so...

I'm not sure where `in-directory` was defined, but it's missing
too. Using a temp buffer seems more emacs-ish anyway.

$